server/requirements.txt
numpy
pycrypto
psycopg2
Flask